fix: resolve bugs from code review (keypoint, NMS, MPP, thread-safety)
- fix(infer): keypoint decode double-stride bug in C++ and Python
  anchor_cx is already in pixel coords; remove extra multiply by stride
  affects pose model keypoint output (detect mode unaffected)

- fix(capture): MppSource EOF flush now sends proper EOS packet
  mpp_packet_set_eos required to drain decoder internal buffer;
  null-data packet without EOS caused last frames to be lost

- fix(post): OpenMP parallel for NMS break UB and box coverage gap
  break inside #pragma omp parallel for is undefined behavior;
  refactor to pre-compute aligned_end, handle tail serially outside loop;
  also fix memory_order_relaxed -> acquire for suppression checks

- fix(common): DmaBuf::getRgaHandle lazy init is now thread-safe
  add mutable rga_mutex_ with double-checked locking to prevent
  concurrent importbuffer_fd calls leaking RGA handles

- fix(tools): calib.txt now writes absolute paths
  glob results were relative to CWD; RKNN calibration fails silently
  with relative paths; use Path.resolve() for all entries

- fix(infer): remove static once_flag from RAW decode log
  process-global once_flag silenced logs from all subsequent engines;
  restore per-call logging

- fix(config): replace generator.throw() anti-pattern in validators
  replace with standard def + raise for correct exception propagation

+++ src/infer/rknn/RknnEngine.cpp
@@ -396,12 +396,14 @@ std::vector<Detection> RknnEngine::inferPreprocessed(
          preprocessed_image.rows);
     return {};
   }
+  // Impl 字段在 init() 后不再变更，通过持有 shared_ptr 保证对象存活；
+  // 此处读取无需再持有 state_mutex_（init 与 infer 不允许并发）。
   const int out_n_snapshot = impl->out_n;
   const int out_c_snapshot = impl->out_c;
   const int out_elems_snapshot = impl->out_elems;
-  const int out_n_dims_snapshot = impl->out_attr.n_dims;
-  const int out_dim1_snapshot = impl->out_attr.dims[1];
-  const int out_dim2_snapshot = impl->out_attr.dims[2];
+  const int out_n_dims_snapshot = static_cast<int>(impl->out_attr.n_dims);
+  const int out_dim1_snapshot = static_cast<int>(impl->out_attr.dims[1]);
+  const int out_dim2_snapshot = static_cast<int>(impl->out_attr.dims[2]);
   const uint32_t out_index_snapshot = impl->out_attr.index;
 
   // RKNN 训练/转换常用 RGB 输入；上游传入 BGR，这里统一转换一次。
